Added Deployment Scripts
- Added docker deployments
- Added Domain Layer Modifications
- Corrected the Dockerfile
