+++ SKB.CleanWebApi.sln
@@ -16,6 +16,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\Applicat
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Contracts", "src\Contracts\Contracts.csproj", "{D2705AEB-B238-4E76-ABEA-FF3E23C987D5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\Domain\Domain.csproj", "{230A11F5-7231-4CF8-AFE1-795AE7146B8B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,5 +36,9 @@ Global
 		{D2705AEB-B238-4E76-ABEA-FF3E23C987D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D2705AEB-B238-4E76-ABEA-FF3E23C987D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D2705AEB-B238-4E76-ABEA-FF3E23C987D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{230A11F5-7231-4CF8-AFE1-795AE7146B8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{230A11F5-7231-4CF8-AFE1-795AE7146B8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{230A11F5-7231-4CF8-AFE1-795AE7146B8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{230A11F5-7231-4CF8-AFE1-795AE7146B8B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
