+++ src/CleanWebApi.Http/appsettings.json
@@ -6,5 +6,5 @@
 		}
 	},
   	"AllowedHosts": "*",
-	"OTEL_EXPORTER_OTLP_ENDPOINT": "http://localhost:4317"
+	"OTEL_EXPORTER_OTLP_ENDPOINT": "http://skb-otel-collector:4317"
 }
