+++ src/Contracts/GetWeatherForecast/GetWeatherForecastQueryDto.cs
@@ -1,18 +1,6 @@
-﻿namespace SKB.App.Contracts.GetWeatherForecast;
+﻿using SKB.App.Domain.GetWeatherForecast;
 
-/// <summary>
-/// Data transfer object for WeatherForecast information
-/// </summary>
-public record GetWeatherForecastTransmissionData(
-	DateOnly Date,
-	int TemperatureC,
-	string? Summary)
-{
-	/// <summary>
-	/// Temperature in F
-	/// </summary>
-	public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
+namespace SKB.App.Contracts.GetWeatherForecast;
 
 /// <summary>
 /// Singular Entropy Dto for GetWeatherForecastQuery
